one-to-many

+++ Controllers/PetsController.cs
@@ -1,8 +1,11 @@
 using System;
+using System.Linq;
 using ClientNotifications;
 using Demo.Models;
 using Demo.Repositories;
 using Demo.ViewModels;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using static ClientNotifications.Helpers.NotificationHelper;
 
@@ -11,10 +14,12 @@ namespace Demo.Controllers
     public class PetsController : Controller
     {
         private readonly IPetRepository _repo;
+        private readonly UserManager<AppUser> _userManager;
 
-        public PetsController(IPetRepository repo)
+        public PetsController(IPetRepository repo,UserManager<AppUser> userManager)
         {
             _repo = repo;
+            _userManager = userManager;
         }
         public IActionResult Index(string search=null)
         {
@@ -25,6 +30,11 @@ namespace Demo.Controllers
             }
             return View(_repo.GetAllPets());
         }
+        public IActionResult MyPets()
+        {
+            var pets = _repo.GetPetsByUserId(GetUserId());
+            return View(pets);
+        }
         public IActionResult Details(int id)
         {
             var pet = _repo.GetSinglePet(id);
@@ -34,13 +44,14 @@ namespace Demo.Controllers
             }
             return View(pet);
         }
-
+        [Authorize]
         [HttpGet]
         public IActionResult New()
         {
             var pet = new NewPetVM();
             return View(pet);
         }
+        [Authorize]
         [HttpPost]
         public IActionResult New(NewPetVM model)
         {
@@ -52,12 +63,14 @@ namespace Demo.Controllers
             {
                 Name = model.Name,
                 Age = model.Age,
-                Color = model.Color
+                Color = model.Color,
+                UserId = GetUserId()
             };
                 _repo.Create(pet);
 
                 return RedirectToAction(nameof(Index));
         }
+        [Authorize]
         public IActionResult Edit(int id)
         {
             var pet = _repo.GetSinglePet(id);
@@ -65,6 +78,10 @@ namespace Demo.Controllers
             if (pet == null)
             {
                 return NotFound();
+            }
+            if (pet.UserId!=GetUserId())
+            {
+                return RedirectToAction("AccessDenied","Account");
             }
             var model = new EditPetVM
             {
@@ -76,6 +93,7 @@ namespace Demo.Controllers
             };
             return View(model);
         }
+        [Authorize]
         [HttpPost]
         public IActionResult Edit(EditPetVM model)
         {
@@ -92,6 +110,7 @@ namespace Demo.Controllers
                 return RedirectToAction(nameof(Index));
           
         }
+        [Authorize]
         [HttpPost]
         public IActionResult Delete(int id)
         {
@@ -99,6 +118,10 @@ namespace Demo.Controllers
             if (pet == null)
             {
                 return NotFound();
+            }
+            if (pet.UserId != GetUserId())
+            {
+                return RedirectToAction("AccessDenied", "Account");
             }
             _repo.Delete(pet);
             return RedirectToAction(nameof(Index));
@@ -114,6 +137,11 @@ namespace Demo.Controllers
         public IActionResult AutocompleteResult(string search)
         {
             return Json(_repo.SearchPets(search));
+        }
+
+        private string GetUserId()
+        {
+            return _userManager.GetUserId(this.HttpContext.User);
         }
     }
 }
